requirements.txt
python-binance
asyncio
numpy
requests
beautifulsoup4
pandas
datetime
openpyxl
numpy
python-dotenv